requirements.txt
flask
flask_sqlalchemy
flask_jwt_extended
flask_cors
psycopg2